src/requirements.txt
black==v19.10b0
flake8>=3.7.9, <4.0
ipython~=7.0
isort>=4.3.21, <5.0
jupyter~=1.0
jupyter_client>=5.1, < 7.0
kedro==0.16.6
datasets==1.2.1
kedro-viz>=3.0.0,<4.0.0
kedro-mlflow>=0.4.0,<0.5.0
# numpy==1.18.5 ## several known bugs with numpy>1.19
scikit-learn==0.23.2
xgboost>=1.2.0,<1.3.0
pandas>=1.1.0,<1.2.0
matplotlib>=3.3.0,<3.4.0
keras-preprocessing==1.1.2